Add hash for Windows icon generation
This avoid running Inkscape and other tools to generate the icon
if the source SVG file has not changed.

+++ internal/icons/windows/generate-icon.sh
@@ -1,17 +1,26 @@
 #!/usr/bin/env nix-shell
 #! nix-shell -i bash -p inkscape imagemagick
 
-tmp=$(mktemp -d)
+svgInput="./hicolor/scalable/apps/so.libdb.dissent.svg"
+icoOutput="./windows/dissent.ico"
 
-inkscape \
-	-w 256 -h 256 \
-	-o $tmp/logo.png \
-	./hicolor/scalable/apps/so.libdb.dissent.svg
+svgHash=$(sha256sum "$svgInput" | cut -d' ' -f1)
+svgHashFile="./windows/so.libdb.dissent.svg.sha256"
 
-convert \
-	-define icon:auto-resize=256,128,96,64,48,32,16 $tmp/logo.png \
-	-colors 256 $tmp/logo.ico
+if [[ ! -f $svgHashFile || $(< $svgHashFile) != $svgHash ]]; then
+	tmp=$(mktemp -d)
 
-mv $tmp/logo.ico ./windows/dissent.ico
+	inkscape \
+		-w 256 -h 256 \
+		-o $tmp/logo.png \
+		"$svgInput"
+	
+	convert \
+		-define icon:auto-resize=256,128,96,64,48,32,16 $tmp/logo.png \
+		-colors 256 $tmp/logo.ico
+	
+	mv $tmp/logo.ico "$icoOutput"
+	rm -r $tmp
 
-rm -r $tmp
+	echo $svgHash > $svgHashFile
+fi
